Logica batalha automatica

+++ include/digicomm.h
@@ -4,9 +4,9 @@
 #include <stdint.h>
 
 #define DIGICOMM_OK            0
-#define DIGICOMM_ERROR_POLLING (DIGICOMM_OK - 1)
-#define DIGICOMM_ERROR_READING (DIGICOMM_OK - 2)
-#define DIGICOMM_ERROR_WRITING (DIGICOMM_OK - 3)
+#define DIGICOMM_ERROR_POLLING (uint8_t)(DIGICOMM_OK - 1)
+#define DIGICOMM_ERROR_READING (uint8_t)(DIGICOMM_OK - 2)
+#define DIGICOMM_ERROR_WRITING (uint8_t)(DIGICOMM_OK - 3)
 
 uint16_t DIGICOMM_setup();
